fix(APISIMP-DQR-LIST-IN-MEMORY-PAGING): use long skip to prevent CodeQL overflow
int skip = page * pageSize overflows when page has no @Max bound; cast
page to long before multiply so CodeQL user-controlled-arithmetic is satisfied.

+++ backend/src/main/java/de/dlr/shepard/v2/quality/daos/DataQualityRequirementDAO.java
@@ -49,13 +49,13 @@ public class DataQualityRequirementDAO extends GenericDAO<DataQualityRequirement
    * @param limit           maximum rows to return
    * @return ordered slice (by appId ascending); empty list when none in range
    */
-  public List<DataQualityRequirement> findByCollectionAppId(String collectionAppId, int skip, int limit) {
+  public List<DataQualityRequirement> findByCollectionAppId(String collectionAppId, long skip, int limit) {
     String query =
       "MATCH (d:DataQualityRequirement)-[:APPLIES_TO]->(c:Collection {appId: $collectionAppId}) " +
       "RETURN d ORDER BY d.appId ASC SKIP $skip LIMIT $limit";
     Map<String, Object> params = new HashMap<>();
     params.put("collectionAppId", collectionAppId);
-    params.put("skip", (long) skip);
+    params.put("skip", skip);
     params.put("limit", (long) limit);
     List<DataQualityRequirement> result = new ArrayList<>();
     for (var d : findByQuery(query, params)) {
